+++ wiki/Changelog.md
@@ -1,3 +1,5 @@
+### 3.4.3 (11 November 2020) ###
+
 Character encoding
   * Fixed locale setup interworking with bash startup profile (~#1050).
   * Handling empty startup locale to be consistent with system-derived shell locale (~#1050).